pyproject.toml
[tool.poetry]
name = "bill-parser-engine"
version = "0.1.0"
description = "A tool to parse and translate French legislative bills into clear, digestible text"
authors = ["Du Phan"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
pydantic = "^2.11.5"
httpx = "^0.28.1"
python-dotenv = "^1.1.0"
loguru = "^0.7.3"
mistralai = "^1.8.1"
pytest = "^8.3.5"
requests = "^2.32.3"
# pylegifrance removed; French articles are read from local store


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
